pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = ""
authors = ["Povarenskiy <Povarenskiy.rb@yandex.ru>"]
readme = "README.md"
packages = [{include = "notification_service"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
celery = "^5.2.7"
redis = "^4.5.1"
requests = "^2.28.2"
pytz = "^2022.7.1"
gevent = "^22.10.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
